+++ cpp/algorithms/sorting/heap/lib-heap.hpp
@@ -18,7 +18,7 @@ size_t Parent(const size_t &index);
 size_t Left(const size_t &index);
 size_t Right(const size_t &index);
 
-void MaxHeapify(std::vector<int> &array, size_t thisIndex, const int &heapSize);
+void MaxHeapify(std::vector<int> &array, size_t thisIndex, const size_t &heapSize);
 
 void BuildMaxHeap(std::vector<int> &array);
